refactor(core): rename ThetaGPManger to ThetaGamepad with singleton, add MempoolManager init
- Rename ThetaGPManger to ThetaGamepad, add getInstance() singleton
- Add MempoolManager init before peripheral setup in ThetaGamepad::Setup()
- Move bus allocBuf after descriptor init in UartBus constructors
- Remove default buffer sizes from Bus base class, add explicit _pTxBufSize/_pRxBufSize
- Make UartBus write() copy through _pTxBuf before HAL transmit
- Rename BusMemPool to s_BusMempool, remove _initialized guard
- Clean up scheduler includes and namespace usage
- Add LOG_DEBUG to peripheralsmgr init

Model: claude-3.5-sonnet
Agent: opencode

+++ src/drivers/peripherals/bus/bus.cpp
@@ -20,7 +20,7 @@ Bus::Bus()
 
 Bus::~Bus() { freeBuf(); }
 
-void Bus::allocBuf(size_t txSize, size_t rxSize) {
+void Bus::allocBuf(uint32_t txSize, uint32_t rxSize) {
   if (_pTxBuf == nullptr) {
     _pTxBuf = static_cast<uint8_t *>(_busMem.allocTxBuffer(txSize));
     if (_pTxBuf != nullptr) {
